refactor(tools/mcp-conformance): hoist responseText/track/structured into _runner.cjs
Three SDK-envelope-shape helpers survived the _runner.cjs factoring as
copy-pasted bodies across the conformance test files. None pins a per-test
contract — all are pure SDK-envelope mechanics feeding helpers the runner
already exports. Hoist them into a shared 'envelope readers / coverage tracker'
section and delete the local copies:

- responseText (4 copies, functionally identical) — concatenate the .text of
  every callTool content block.
- track (2 byte-identical copies) — the callTool coverage Proxy. The shared
  form owns its CALLED set and returns { client, called } so the call sites
  destructure rather than reaching a module-scope Set.
- structured (2 copies) — decodeDedupEnvelope unwrap; the decodeDedupEnvelope
  import moves with it, dropped from story + flag-gates. The null-safe
  ((resp && …) || {}) form is kept (a strict superset of story's no-guard
  copy, so behaviour-preserving for both).

Zero behaviour change. The per-schema validators (assertOverflowBody /
assertProgressParams / assertDropped / dispatchStamp …) are left in place —
those are genuinely-distinct per-wire-surface conformance assertions.
